fix: move encrypted icon to excalidraw-app add separate animation for `renderFooter` prop (#3577)
* fix: move encrypted icon to excalidraw-app

* use grid & separate animation for custom footer

* update docs

* fix
